+++ packages/terra-application-navigation/src/terra-dev-site/test/application-navigation/Title.test.jsx
@@ -1,5 +1,4 @@
 import React from 'react';
-import ApplicationBase from 'terra-application/lib/application-base';
 import ApplicationNavigation from '../../../ApplicationNavigation';
 import FixedHeight from '../common/FixedHeight';
 import Logo from '../common/Logo';
@@ -30,13 +29,13 @@ const titleConfigHideTitle = {
 
 // TODO: remove terra-application after it is incorporated into dev-site for themes or themes are co-located
 const TitleTest = () => (
-  <ApplicationBase locale="en">
+  <>
     <FixedHeight><ApplicationNavigation titleConfig={titleConfig} /></FixedHeight>
     <FixedHeight><ApplicationNavigation titleConfig={titleConfigHeadLine} /></FixedHeight>
     <FixedHeight><ApplicationNavigation titleConfig={titleConfigSubline} /></FixedHeight>
     <FixedHeight><ApplicationNavigation titleConfig={titleConfigElement} /></FixedHeight>
     <FixedHeight><ApplicationNavigation titleConfig={titleConfigHideTitle} /></FixedHeight>
-  </ApplicationBase>
+  </>
 );
 
 export default TitleTest;